+++ client/src/language/xml/XmlDiagnostics.ts
@@ -20,6 +20,7 @@ import * as xmlChecker from 'xmlChecker';
 import { IDiagnose, ui5tsglobal } from "../../extension";
 import * as extension from '../../extension';
 import { File } from "../../helpers/filehandler";
+import { getController } from "../searchFunctions";
 
 export interface I18nLabel {
     text: string;
@@ -96,7 +97,7 @@ export class I18nDiagnosticProvider implements IDiagnose {
         this.diagnosticCollection.set(document.uri, this.diagi18n(document));
     }
 
-    public diagi18n(document: TextDocument): Diagnostic[] {
+    private diagi18n(document: TextDocument): Diagnostic[] {
         try {
             const text = document.getText();
             const i18nreg = new RegExp("\"\\s*?{\\s*?" + ui5tsglobal.config["lang.i18n.modelname"] + "\\s*?>\\s*?(.*?)\\s*?}\\s*?\"", "g");
@@ -119,17 +120,36 @@ export class I18nDiagnosticProvider implements IDiagnose {
         }
     }
 
-    public getRange(document: TextDocument, startIndex: number, length: number): Range {
+    private getRange(document: TextDocument, startIndex: number, length: number): Range {
         return new Range(document.positionAt(startIndex), document.positionAt(startIndex + length));
     }
 }
 
 export class ControllerDiagnosticsProvider implements IDiagnose {
     constructor(public diagnosticCollection: DiagnosticCollection) {
-
     }
 
-    public diagnose(document: TextDocument) {
-        throw new Error('Not implemented yet.');
+    public async diagnose(document: TextDocument) {
+        this.diagnosticCollection.delete(document.uri);
+        this.diagnosticCollection.set(document.uri, await this.diagController(document));
+    }
+
+    private async diagController(document: TextDocument): Promise<Diagnostic[]> {
+        const text = document.getText();
+        // 1: text before controller name
+        // 2: quotes
+        // 2: controller name as namespace
+        const mController = text.match(/(\bcontrollerName\b\s*=\s*("|'))([\w\.]+?)\2/);
+        if (mController) {
+            const controllers = await getController(document);
+            if ((!controllers) || controllers.length < 1) {
+                return [new Diagnostic(this.getRange(document, mController.index + mController[1].length, mController[3].length), "Could not find corresponding controller", vscode.DiagnosticSeverity[ui5tsglobal.config["lang.xml.linter.controller"]])];
+            }
+        }
+        return [];
+    }
+
+    private getRange(document: TextDocument, startIndex: number, length: number): Range {
+        return new Range(document.positionAt(startIndex), document.positionAt(startIndex + length));
     }
 }